+++ c/sungyonglee/eleven/15/quiz15.h
@@ -1,6 +1,6 @@
 #ifndef __QUIZ15_H__
 #define __QUIZ15_H__
 
-float quiz15(int number1, float number2);
+int quiz15(int number1, float number2);
 
 #endif
